render all citations together, to allow author-date disambiguation: https://github.com/metanorma/metanorma-ieee/issues/103
