+++ lib/isodoc/presentation_function/section.rb
@@ -1,4 +1,4 @@
-require_relative "../../relaton/render/general"
+require_relative "refs"
 
 module IsoDoc
   class PresentationXMLConvert < ::IsoDoc::Convert
@@ -74,85 +74,6 @@ module IsoDoc
     def term1(elem)
       lbl = @xrefs.get[elem["id"]][:label] or return
       prefix_name(elem, "", "#{lbl}#{clausedelim}", "name")
-    end
-
-    def references(docxml)
-      bibliography_bibitem_number(docxml)
-      docxml.xpath(ns("//references/bibitem")).each do |x|
-        bibitem(x)
-      end
-      docxml.xpath(ns("//references[bibitem/@hidden = 'true']")).each do |x|
-        x.at(ns("./bibitem[not(@hidden = 'true')]")) and next
-        x["hidden"] = "true"
-      end
-      @xrefs.parse_inclusions(refs: true).parse(docxml)
-    end
-
-    def bibitem(xml)
-      @xrefs.klass.implicit_reference(xml) and
-        xml["hidden"] = "true"
-      bibrender(xml)
-    end
-
-    def bibrender(xml)
-      if f = xml.at(ns("./formattedref"))
-        bibrender_formattedref(f, xml)
-      else bibrender_relaton(xml)
-      end
-    end
-
-    def bibrender_formattedref(formattedref, xml)
-      code = render_identifier(bibitem_ref_code(xml))
-      (code[:sdo] && xml["suppress_identifier"] != "true") and
-        formattedref << " [#{code[:sdo]}] "
-    end
-
-    def bibrender_relaton(xml)
-      bib = xml.dup
-      bib["suppress_identifier"] == true and
-        bib.xpath(ns("./docidentifier")).each(&:remove)
-      xml.children =
-        "#{bibrenderer.render(bib.to_xml)}"\
-        "#{xml.xpath(ns('./docidentifier | ./uri | ./note')).to_xml}"
-    end
-
-    def bibrenderer
-      ::Relaton::Render::IsoDoc::General.new(language: @lang)
-    end
-
-    def bibliography_bibitem_number_skip(bibitem)
-      @xrefs.klass.implicit_reference(bibitem) ||
-        bibitem.at(ns(".//docidentifier[@type = 'metanorma']")) ||
-        bibitem.at(ns(".//docidentifier[@type = 'metanorma-ordinal']")) ||
-        bibitem["hidden"] == "true" || bibitem.parent["hidden"] == "true"
-    end
-
-    def bibliography_bibitem_number(docxml)
-      i = 0
-      docxml.xpath(ns("//references[@normative = 'false']/bibitem")).each do |b|
-        i = bibliography_bibitem_number1(b, i)
-      end
-      @xrefs.references docxml
-    end
-
-    def bibliography_bibitem_number1(bibitem, idx)
-      if mn = bibitem.at(ns(".//docidentifier[@type = 'metanorma']"))
-        /^\[?\d\]?$/.match?(mn&.text) and
-          idx = mn.text.sub(/^\[/, "").sub(/\]$/, "").to_i
-      end
-      unless bibliography_bibitem_number_skip(bibitem)
-
-        idx += 1
-        bibitem.at(ns(".//docidentifier")).previous =
-          "<docidentifier type='metanorma-ordinal'>[#{idx}]</docidentifier>"
-      end
-      idx
-    end
-
-    def docid_prefixes(docxml)
-      docxml.xpath(ns("//references/bibitem/docidentifier")).each do |i|
-        i.children = @xrefs.klass.docid_prefix(i["type"], i.text)
-      end
     end
 
     def index(docxml)
